requirements.txt
nltk==3.3
numpy==1.15.1
tqdm==4.26.0
allennlp==0.8.5
pytorch-ignite==0.2.0
pytorch_transformers==1.1.0
torch==1.1.0
PyYAML==5.1.2
